Quiet warning

+++ intern/cycles/util/util_md5.cpp
@@ -367,7 +367,7 @@ string MD5Hash::get_hex()
 	finish(digest);
 
 	for(int i = 0; i < 16; i++)
-		sprintf(buf + i*2, "%02X", digest[i]);
+		sprintf(buf + i*2, "%02X", (unsigned int)digest[i]);
 	buf[sizeof(buf)-1] = '\0';
 	
 	return string(buf);
